Split off CSS and renamed the rewriter files
